integrates/06.langchain-integration/requirements.txt
# Python 依赖包
langchain>=0.1.0
langchain-openai>=0.0.5
openfga-sdk>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0

# 可选依赖
# 如果需要使用其他 LLM
# langchain-anthropic>=0.1.0
# langchain-google-genai>=0.0.5

# 开发依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
